desloppify: decompose process.rs into focused submodules
commands/process.rs had grown to 2748 LOC / 32 functions, hosting the
per-file processing pipeline, plan-lifecycle event dispatch, three
safeguards (size/duration/disk), outcome reconstruction, and file-
transition recording — all inside what was nominally a CLI dispatcher
module. Five blind reviewers flagged it across six dimensions.

Convert to a subdirectory with focused modules:

  process/
  ├── mod.rs (1498)      thin dispatcher: run() + helpers
  ├── pipeline.rs (663)  per-file processing loop
  ├── dispatch.rs (158)  PlanDispatcher + log_plugin_errors pairing
  ├── safeguards.rs (451) size/duration/disk checks + shared
  │                       record_safeguard_failure helper
  ├── plan_outcome.rs (76) PlanOutcome::from_event_result
  └── transitions.rs (109) FileTransitionContext + record_file_transition

Key cleanups surfaced by the decomposition:

- `record_file_transition` loses its 10-arg signature (and the
  clippy::too_many_arguments allow). Callers now build a
  FileTransitionContext struct, matching the ProcessContext /
  RunResultsContext pattern already used throughout mod.rs.

- EventResult → PlanOutcome reconstruction is done ONCE in
  plan_outcome::PlanOutcome::from_event_result instead of three
  separate scans at the caller.

- Every kernel.dispatch(event) site in the former process.rs paired
  with log_plugin_errors — these ~8 copy-paste sites are now
  dispatch::dispatch_and_log / PlanDispatcher calls.

- The PlanExecuting → PlanCreated → PlanCompleted/Skipped/Failed
  ordering contract is documented in one place (dispatch.rs module
  doc) and enforced by PlanDispatcher's method order.

- The three safeguards (check_size_increase, check_duration_shrink,
  check_disk_space) share record_safeguard_failure() to fold the
  repeated PlanFailed dispatch + record_phase_stat(Failed) +
  record_failure_transition triplet.

Also fixes a pre-existing functional test that was failing after the
earlier auth_token redaction change:
  config_set_and_get now round-trips through default_policy (non-
  secret) and a new config_get_auth_token_is_redacted test locks in
  the redaction contract for the secret key.

Resolves these review findings:
  high_level_elegance::process_cmd_absorbing_pipeline
  mid_level_elegance::plugin_error_logging_boilerplate_at_dispatch_sites
  mid_level_elegance::plan_lifecycle_dispatch_contract_scattered_across_helpers
  mid_level_elegance::plan_outcome_reconstruction_from_eventresult_bag
  low_level_elegance::duplicated_safeguard_failure_epilogue
  low_level_elegance::record_file_transition_10_args_bypasses_local_convention

1725 workspace tests + 104 functional tests pass. Default clippy clean.
